refactor(components): let HP carry fractions
Damage-over-time deals halves of a point. An int HP bar rounds those to
nothing, so a burn tick would be silently free. Hp, MaxHp, TakeDamage and
the Damaged signal all widen to double; int callers are unaffected.

+++ scripts/components/DamageFlashComponent.cs
@@ -48,7 +48,7 @@ public partial class DamageFlashComponent : Node
             Health.Damaged -= OnDamaged;
     }
 
-    private void OnDamaged(int amount, int hp)
+    private void OnDamaged(double amount, double hp)
     {
         _tween?.Kill();
         _tween = CreateTween().SetParallel(true);
